Windows mouse handling affected by QuickEdit mode. Added forced disabling of QuickEdit mode when enabling mouse handling. Moved Alertbox into separate module.
