+++ src/terminal/wincon/mouse.rs
@@ -10,7 +10,7 @@ const MOUSE_MODE: u32 = 0x0010 | 0x0080 | 0x0008;
 pub fn enable_mouse_mode() -> Result<()> {
     let handle = Handle::conin()?;
     let mode = handle.get_mode()?;
-    let mouse_mode = mode | MOUSE_MODE;
+    let mouse_mode = (mode | MOUSE_MODE) & !0x0040;
     handle.set_mode(&mouse_mode)?;
     Ok(())
 }
@@ -18,7 +18,7 @@ pub fn enable_mouse_mode() -> Result<()> {
 pub fn disable_mouse_mode() -> Result<()> {
     let handle = Handle::conin()?;
     let mode = handle.get_mode()?;
-    let mouse_mode = mode & !MOUSE_MODE;
+    let mouse_mode = (mode & !MOUSE_MODE) | 0x0040;
     handle.set_mode(&mouse_mode)?;
     Ok(())
 }
